backend/requirements.txt
flask
flask-cors
mariadb==1.1.*
pyjwt
